creating SearchBox
